build: Do not build with Remotery by default (#466)
#### Reference Issues/PRs

Extracted from #449.

#### What does this implement/fix? Explain your changes.

Remotery seems to be used as a performance analysis tool for
development.
Remotery might not be needed when it is released to users.

This PR proposes to add a `BUILD_WITH_REMOTERY` CMake option (`OFF` by
default) to build ArcticDB with Remotery if needed. It reuses the
existing macro.

#### Any other comments?

ArcticDB for conda-forge will need not be built with Remotery. 
Should we turn `BUILD_WITH_REMOTERY` to `ON` by default?

What do you think, @mehertz and @GemDot?

---------

Signed-off-by: Julien Jerphanion <git@jjerphan.xyz>
Co-authored-by: Qi Chen <Qi.Chen@man.com>