Adapt denornalization
The normalizaton and denormalization of empty Series and empty
DataFrame is tied and has to slightly change with pandas 2.0.

Thus, we can't satisfy the original test_fallback_to_pickle (which
test empty DataFrame normalization and denormalization) and
test_write_no_rows (which test empty Series' normalization and
denormalization).

The current choice of Pandas 2.0 allows for DataFrame and Series not to
be pickled to be stored, which is currently the case with Pandas 1.0.

Signed-off-by: Julien Jerphanion <git@jjerphan.xyz>