Fix flaky test_filter_numeric_isnotin_unsigned (#496)
The cause was type_arithmetic_promoted_type would return int64 as the common type for uint64 and any signed int. When we then do the static_cast<WideType>(*ptr++) before calling the Is[Not]InOperator, the uint64 is converted to int64 and the special overloads in the Operators are never used.

+ Ability to only_test_encoding_version_v1 in a test/class/module