Pass `ReadOptions` rather than individual options; Fix `as_of` with timestamp reading entire version chain (#596)
Using AWS replication it's possible that the version key associated with
the most recent reference key update might not yet be available. In
order to mitigate this, we store both the first- and second-most-recent
version keys in the reference key structure, and in the event that the
first one is not readable use the second to iterate the version chain.

NOTE: In the process of making this change it became clear that loading
the version map as_of a particular time loads the whole map without
early termination. This is problematic for the above change because the
assumption that only a load of the latest version should terminate the
version chain following after the reference key is broken, therefore
it's necessary to have correct early termination when searching the
version list by timestamp.

This is now the major change in this PR. The promotion of the
penultimate key to the reference segment in the version list is still in
the code, but is currently disabled to mitigate technical risk (as it
changes the stored format and might be difficult to back out), and
whilst we decide whether the API is optimal.

This has been done because correct early termination of the version list
traversal by timestamp needs to be fixed sooner rather than later, and
disentangling the changes would be difficult and slow

---------

Co-authored-by: mhertz <Matthew.Hertz@man.com>