Refactor underlying segment write scheduling  (#532)
The current version of batch_write which is called by a normal
lib.write() to write the data segments collects on the encoded segments
so that it can compare them to the dedup map. This in unecessary as the
items in the dedup map already exist, and the by-hand batching
functionality can be replaced with a folly::window to reduce tail
latency.

Enqueuing the (possible) writes to storage immediately after the
compress and check for duplicates results in a 7x speedup in the
high_entropy (i.e bandwidth-constrained) test case in
test_stress_generated, which is a public benchmark from Wes McKinney's
blog.