Fixes IFNDR issue due to mismatching inline/non-inline functions depending on the translation unit (fixes #943)
This fixes an IFNDR (Ill-Formed, No Diagnostic Required) issue explained
in #943 : storage implementations are defined as inline in one
cpp/translation unit and not-inline in others, leading to potential
symbols not emitted.
While the issue is currently only visible using Windows/MSVC on my
machine, the issue is actually more general and needs to be fixed to
reason correctly about the whole build, whatever the target platform or
build toolchain or flags etc.

The solution used here is to move inline code that should not be into
cpp files.