Remove Black and pre-commit setup
Black does not reliably work in the CI. Example:

https://github.com/man-group/ArcticDB/issues/549

pre-commit has issues on Windows, as it blows up when it encounters
symlinks.

Since this linting is interfering with getting actual work merged, this
PR proposes removing it for now.

We can set up black formatting again later with a simple bash script
that just runs on Linux. Ticket for that:

https://github.com/man-group/ArcticDB/issues/971