Bugfix 1035: Fix grouping on float columns in linux conda (#1261)
Closes #1035 

There is still an issue with non-deterministic hashing in the linux
conda build, #1268 opened to investigate this further.
In addition:

- We were previously dynamically linking `xxhash` in the conda builds,
despite using macros designed for static linking. Now it is linked in
statically.
- We were not pinning the `xxhash` version in the conda builds, and so
were on the latest version `0.8.2`. This is now pinned to `0.8.2`. This
pin will presumably need to be replicated in the feedstock.
- We were using `xxhash.c` and `xxhash.h` files copied from version
`0.6.5` of the project. We now pull in this dependency using `vcpkg`,
pinned to version `0.8.2`.
-
https://fastcompression.blogspot.com/2018/03/xxhash-for-small-keys-impressive-power.html
claims that for small keys, if the size argument to the `xxhash` hashing
functions is a compile-time constant, then performance improvements can
be observed, so an overload of `hash` was added to do this. This is the
method used for hashing aggregations, where the keys are at most 8 bytes
long for non-string columns.