#447 LMDB Exceptions Normalization (#1285)
Implementation of lmdb storage specific exceptions. This PR is a
refactor of #584 which was not merged because it was too complex.

This PR also implements a framework for testing storage exceptions.
`GenericStorageTest` class allows testing of common exceptions which are
triggered among all the storages. Some examples of these exceptions are
`KeyNotFoundException` if you read an in-existent key,
`DuplicateKeyException` if you attempt to overwrite a key that already
exists.

`lmdb::map_full_error` is an example of exception specific to LMDB which
is triggered when the remaining map size is not enough to write new
keys/values.