Abstract AzureClientWrapper out of azure_storage.cpp (#1315)
This is a refactor which should not modify behavior.

The refactor abstracts away building the requests and parsing the
responses from Azure behind an abstract class `AzureClientWrapper`. The
refactor is written similar to how #1274 is for S3.

`RealAzureClient` (derives `AzureClientWrapper`) is implemented by
moving the Azure Client logic from `azure_storage.cpp`.
`AzureClientWrapper` will then be used to add `MockAzureClient` class
which would be used to simulate azure erros.