Bugfix 1334: optimise version ref key access (#1345)
#### Reference Issues/PRs
Fixes #1334 

#### What does this implement or fix?
Main functional change is to `read_symbol_ref` in `version_utils.hpp`.
There was an old style of version chain entry point which used a
`VERSION` key of type reference (rather than atom), as opposed to a
`VERSION_REF` key. The old implementation did a HEAD request on both of
these keys, and then read the one that exists. This introduced 1 or 2
tiny IOs, when the same information could be obtained by attempting to
GET the relevant keys, and handling the key not existing as appropriate.

Now reading and deleting operations only ever read the version ref key
once. Modifying operations still read the key twice, to account for the
fact that other changes could have been made to the symbol between the
latest version being identified, and being ready to write the new
version ref key back.
The only exception is if `write_metadata` is called with a symbol name
that has never existed. This case should be rare, and the code change to
reduce this to 2 reads in line with other modifying operations would
make the general `write` path harder to read.

The remaining changes are to strip out attempts to read
`VERSION_JOURNAL` keys, as these should all have been replaced with
`VERSION_REF` keys by the (also now deleted) `rewrite_old_journal_keys`
a long time go.