Docs: Increase CSS max-wdith and build docs from a branch (#1363)
Currently, no matter how wide the screen gets, the middle section of the
[website](https://arcticdb-docs.pages.dev/4.3.0/api/arctic/), which
contains the main documentation, always remains fixed to a maximum of
`61rem`. This PR increases the `max-width` to `75rem` to improve
visibility on larger devices.

The [Arctic documentation](https://docs.arcticdb.io/4.3.0/api/arctic/)
page contains tables in docstrings, these tables are not rendered by
mkdocs as it doesn't support tables in docstrings. This PR creates a
separate `arctic_uri.md` markdown which contains the tables and links
the Arctic documentation page to the `arctic_uri.md`.

For testing that the update works as expected, we need to build the docs
from this branch. For this purpose, the docs build script is updated.
Previously, the only way to build the docs from a branch other than
master was to change a docs tag of a released version (e.g.
`v4.2.1-docs`) to the head of the target branch. This is not ideal as
the tag would point to docs which may not be consistent with the version
released. Previously, if the tag is not specified the docs were always
built from master. Now the docs would be built from the selected branch
(which can be any branch including the master).

Please view the effect of the CSS update at the following link:
https://a953c3c5.arcticdb-docs.pages.dev/dev/api/arctic/.