1 year and 1k stars readme banner (#1425)
This pull request introduces a temporary change to the README.md to
celebrate the 1-year anniversary of ArcticDB. A festive banner has been
added to the top of the README to highlight the milestone of ArcticDB
reaching one year in the open-source community and achieving 1,000
GitHub stars.

The banner features a congratulatory message and is accompanied by a
custom image to visually mark the occasion. It is intended to be
displayed for a period of 1 to 2 weeks before being removed.