maint: Remove use of folly/system/ThreadName.h (#1446)
#### Reference Issues/PRs

Fixes one item of the folly replacement plan, issue #1412.

#### What does this implement or fix?

This removes use of `folly/system/ThreadName.h`. It was used in two
places to obtain a `std::string` name of a thread, and here is replaced
with conversion of unique thread ID to a string using `fmt` via
```c++
fmt::format("{}", arcticdb::get_thread_id())
```

#### Any other comments?

This is used in two places, the first in the termination handler in
`python_module.cpp` which I have manually tested. The second use is in
the [Remotery](https://github.com/Celtoys/Remotery) which, as far as I
can tell, is not tested in CI so I am not sure of the status of the
Remotery support.

Signed-off-by: Ian Thomas <ianthomas23@gmail.com>