Support modifying library options, and introduce "enterprise" library options (#1457)
Provide an API to modify library options.

Provide APIs to view and modify "enterprise" library options, which are
options that should only be used by enterprise users.

There is some code churn in here:

- I pushed most of get_library_config up to the `arctic_library_adapter`
as there was so much duplication between adapters for it. I've replaced
it with `add_library_to_env` in the specializations.
- I've pulled out the library management type tests from
`test_arctic.py` to a new file, as `test_arctic.py` is enormous and
covers too much different functionality.

89f45b03a0da04e2927d77095922b76f794d392d fixes some test setup issues
that I was having - my setup was looking for an `eu-west-2` test s3
mapping, presumably due to some config on my machine that I couldn't
track down, and a `brotli` version clash.