Fix SYMBOL_LIST Segment index out of bounds bug
The check `row < row_count()` in `string_at` was problematic for symbol
list entries because they have the exception that it is possible for a
column to have more rows than segment rows.

That exception is because of the nature of the symbol list refactor in
4.2.0 which needed to remain backwards compatible with old readers.

This change bypasses the check for symbol list entries and directly
calls string_at or scalar_at of the Column.