Revert destructor speed regression from #1152
We noticed a speed regression in the version_chain benchmarks arising
from #1152.

Profiling showed that the majority of the slowdown was because the
`~SegmentInMemory` destructor became a lot slower. And all of the
slowdown was because of the recurring `ConfigsMap::get_int` inside
`maybe_trim`.

We fix this by again using a `static` variable for getting the config
value.

I've verified that there aren't other similar regressions from that PR.