Bugfix 1615: Fix segfault in benchmark_clause.cpp (#1616)
#### Reference Issues/PRs
Closes #1615 

After the refactor to speed up hash-based grouping, the C++ grouping
benchmark needs updating to correctly set the last logical row on the
input column, as this is now used to presize the vector mapping rows to
buckets.

Also noticed that some benchmarks did not make sense (e.g. `int8_t`
column with 100,000 unique grouping values), so corrected this as well.