Fix Segment use-after-move when replicating to NFS (#1756)
## Motivation

(this section copied from previous (closed) attempt -
https://github.com/man-group/ArcticDB/pull/1746)

The motivation for the change is to allow `arcticdb-enterprise` to copy
blocks to NFS storages without a use-after-move. I explained this in
https://github.com/man-group/arcticdb-enterprise/pull/139 but to have an
open record:

CopyCompressedInterStoreTask has:

```
                        // Don't bother copying the key segment pair when writing to the final target
                        if (it == std::prev(target_stores_.end())) {
                            (*it)->write_compressed_sync(std::move(key_segment_pair));
                        } else {
                            auto key_segment_pair_copy = key_segment_pair;
                            (*it)->write_compressed_sync(std::move(key_segment_pair_copy));
                        }
```

KeySegmentPair has a shared_ptr to a KeySegmentPair, which we can think
of here as just a `Segment`.

Therefore the old `key_segment_pair_copy` is shallow, the underlying
Segment is the same. But the segment eventually gets passed as an rvalue
reference further down the stack.

In `do_write_impl` we call `put_object` which calls `serialize_header`.

This modifies the segment in place and passes that buffer to the AWS
SDK.

In the `NfsBackedStorage` we have:

```
void NfsBackedStorage::do_write(Composite<KeySegmentPair>&& kvs) {
    auto enc = kvs.transform([] (auto&& key_seg) {
      return KeySegmentPair{encode_object_id(key_seg.variant_key()), std::move(key_seg.segment())};
    });
    s3::detail::do_write_impl(std::move(enc), root_folder_, bucket_name_, *s3_client_, NfsBucketizer{});
}
```

where the segment gets moved from.

Subsequent attempts to use the segment (eg copying on to the next store)
then fail.

https://github.com/man-group/arcticdb-enterprise/pull/139 fixed this
issue by cloning the segment, but this approach avoids the (expensive)
clone.

## Logical Change

Copy the `KeySegmentPair`'s pointer to the `Segment` in
`nfs_backed_storage.cpp` rather than moving from the segment.

## Refactor and Testing

### Copy Task

Move the CopyCompressedInterStoreTask down to ArcticDB from
arcticdb-enterprise. Add a test for it on NFS storage. I've verified
that the tests in this commit fail without the refactor in the HEAD~1
commit. The only changes to `CopyCompressedInterstoreTask` from
enterprise are:

- Pass the `KeySegmentPair` by value in to `write_compressed{_sync}`.
The `KeySegmentPair` is cheap to copy (especially considering we are
about to copy an object across storages, likely with a network hop).
- We have adopted the new `set_key` API of `KeySegmentPair`:
```
if (key_to_write_.has_value()) {
                key_segment_pair.set_key(*key_to_write_);
            }
```
- We have namespaced the `ProcessingResult` struct in to the task

### KeySegmentPair

- Replace methods returning mutable lvalue references to keys with a
`set_key` method.
- Remove the `release_segment` method as it dangerously leaves the
`KeySegmentPair` pointing at a `Segment` object that has been moved
from, and it is not actually necessary.

## Follow up work

The non-const `Segment& KeySegmentPair#segment()` API is still dangerous
and error prone. I have a follow up change to remove it, but that API
change affects very many files and will be best raised separately so
that it doesn't block this fix for replication.

A draft PR showing a proposal for that change is here -
https://github.com/man-group/ArcticDB/pull/1757 .