Add Storage API to check for existence of a key matching a predicate (#1762)
Useful for enterprise (where we want to check if op logs exist on
secondaries as a safety guard).

Remove `test_mongo_storage.cpp` - it just returned immediately and was a
maintenance burden to keep it compiling.

The important API changes are in `storage.hpp`.

We add a new kind of `IterateTypeVisitor` - an `IterateTypePredicate`.
We change the internal `do_iterate_type` functions to take the predicate
instead of the visitor. We change the public `iterate_type` to coerce
its `IterateTypeVisitor` in to a degenerate `IterateTypePredicate` that
never matches anything, so this call still iterates all keys.

Finally we add `test_storage_operations.cpp` to test the new API across
all the supported storage backends.