Fix the sort_index and delete_range wrong version id bug
The problem was that to calculate the next_version_id both `sort_index`
and `delete_range` used the last undeleted version.

This causes issues if we have a version chain like:
```
v0 <- v1 <- v2 <- tombstone(v2)
```

In this case we have to write v3 but these two functions wrote v2.

We fix this by using the same `UpdateInfo` used for other modification
methods.