Add API to remove Prometheus metrics (#1806)
Introduce `void removeMetric(const std::string& name, const Labels&
labels)` to `PrometheusInstance`.

This lets us remove metrics from the Prometheus registry.

The Prometheus SDK has a data model like:

```
Registry
|  The global registry of Prometheus metrics

-> [MetricFamily...] 
|  A list of metric "families". These are metrics that all share the same name, and a set of "base" labels. We create these when we
|  call registerMetric. Since we only call registerMetric once per metric name, we can safely key these in MetricFamilyMap just by
|  metric name.

---->[Metric...]
|      A list of metrics in the family. These metrics are published with the labels of their "family", plus any additional labels (eg 
|      those supplied when calling incrementCounter on a pre-registered metric).
```

We *do not* consider the labels on the `MetricFamily` when removing
metrics because they apply to the entire `PrometheusInstance` and it is
just an unnecessary chore to always include them when removing metrics.

To remove a metric, the SDK forces us to keep a handle on the metric
from when it was created (with `Add`). The SDK's remove method just
compares pointers and silently does not remove metrics unless the
pointers match - there is no deep comparison. Therefore we keep
additional state - a map of metric name and (metric-level) labels to the
created metric. An alternative design would be to push this problem up a
level and force users of the `PrometheusInstance` to maintain state
around their metrics.

Motivation: arcticdb-enterprise#149