Docs 1809: lazy dataframe documentation and example notebook (#1815)
#### Reference Issues/PRs
Closes #1809

#### What does this implement or fix?
Adds documentation and an example notebook for the new `LazyDataFrame`
API for processing operations. Also updated existing example notebooks
to use the `LazyDataFrame` instead of `QueryBuilder`, and the
`QueryBuilder` demo notebook to refer to the `LazyDataFrame` API being
preferred.

Also tweaked the docstrings for `LazyDataFrame` and
`LazyDataFrameCollection` to improve rendering and remove duplication of
`QueryBuilder` docstrings.