Fix a dynamic schema bug
When dynamic schema is used the combined result of all staged
segments had the wrong stream descriptor. It was the result of merging
the current stream descriptor on disk and the merged stream descriptor
for all staged segments. The correct stream descriptor for the
segment is only the result of merging the stream descriptors for all
staged segments. The other option created "empty" columns in the
segment which were not needed. In order to fix that new member to the
pipeline context is added it holds the result from merging the
descriptors of all staged segments.