Mongo destruction segfault fix
When destructing the mongo storage it was possible to destruct the api
instance before destructing the pool.

Moved the instance from the `MongoStorage` to the `MongoClient` to be
kept as a guard around the `mongocxx::pool`. `MongoStorage` didn't need
any sensitive mongocxx objects.

We also disable mongo tests for 4.5.0 as it contains the problematic
segfault.