Use `shell=True` for linux compat tests to fix feedstock runs
The issue with conda feedstock runners is they use some variable for
abspath, which doesn't get correctly expanded by `os.path.expandvar`.

This requires using `shell=True` for linux as well.

This was tested to properly fix compat tests in #1931 combined with
[feedstock pr](https://github.com/conda-forge/arcticdb-feedstock/pull/322).
However it still doesn't work for some of the macos builds which are not skipped correctly.

Still this is a good change for local runs as well and we'll fix the
macos issue in a separate run.