Add a storage test to verify atomic operation support
Adds a `Storage` level check which does on a random non-existent-key:
1. `write_if_none` which should succeed
2. `write_if_none` which should fail due to a precondition failed
3. `remove` to cleanup the key

If the test passes then we deem s3 storage as supporting atomic writes.
The test only catches `AtomicOperationFailedException` and
`NotImplementedByStorageException`. Any other (e.g. network related)
exceptions are propagated up to be handled by users.

Since we now test that the storage raises exactly the
`AtomicOperationFailedException` we only catch that specific exception
when trying to acquire a lock.