Fixes flaky persistant storage lock test
The persistant storage test was deemed flaky because of occasional
`XAmzContentSHA256Mismatch` errors. This was beause of a faulty
multiprocessing sharing of an s3 library. So we reconstruct the `Library` object in each process.

Also upgrades the aws-sdk-cpp (a red herring lead to upgrading the sdk
but didn't fix the issue). Still it's a good change so keeping it.