Chunk up incomplete segments by rows when they are staged (#2117)
This change applies row slicing when incomplete segments are staged, so
by default if we staged an incomplete with 1M rows we now write 10
segments each with 100k rows. The motivation for this is to put some
bound on the memory used when compacting incompletes, as the compaction
step has to hold an entire segment in memory at a time.

I also refactored the code to separate things out better. Previously the
staging code was shared for 3 quite different features:

- Staging an incomplete
- Sorting and staging an incomplete
- Appending an incomplete for tick streaming, forming a linked list of
APPEND_DATA keys

I've separated these out to 3 separate functions now. The sorting case
is quite different to the case without sorting, because to sort we need
to transform the `InputTensorFrame` in to a `Segment` first, whereas
when not sorting we can happily work with the incomplete still as an
`InputTensorFrame`.

The staging an incomplete path is now much more similar to the normal
write path, and we no longer have to worry about `prev_key` and
`next_key` chaining in code paths that never actually use it.

We do not apply column slicing here as that makes it harder for
compaction to infer the schema.

I renamed `append_map.hpp` to `incompletes.hpp` as I think that's more
meaningful to most of us.

Monday: 8066284387