Fixes a use-after-move in `CopyCompressedInterStoreTask`
There was a use-after-move in `CopyCompressedInterStoreTask`. The issue
was highlighted by a failing `Async.CopyCompressedInterStore` C++ test.

The issue is that when we do the `write_compressed_sync` it copies the
`KeySegmentPair` by value (but that only copies the pointers and not the
underlying segment). Then when writing to S3 we move the `Segment` out
of the `shared_ptr`.

This commit fixes this by explicitly doing a deep copy of the
`KeySegmentPair` before doing the `write_compressed_sync`. And after
this the test passes.

The proper fix will be the `KeySegmentPair` refactor.