8234437666 Bound the number of data segments kept live in memory during compaction and sort_merge (#2167)
Monday: 8234437666

This is a simpler alternative to PR #2136 . The idea is to restrict the
number of `write_futures` that we allow to accumulate during
`compact_incomplete` or `sort_and_finalize_staged_data` since each write
task (writing a table data key) holds its segment live in memory, so if
a queue of work builds up we can use an unbounded amount of memory.

This behaviour in memory use is only observable with relatively high
latency storages and large amounts of work. I've simulated it here with
only 1000 incomplete segments, using a faked 3000ms write latency.

We often use `folly::window` to achieve this kind of thing, but that
doesn't fit in well with the `SegmentAggregator` (because the number of
futures we'll end up with is not known ahead of time).

Profiling scripts are available here
https://github.com/man-group/ArcticDB/commit/5e6a3e54fce49b6a9ff01db31b6c20d4d8cd870d

The conclusions from the profiling are:

```
All on my 20 core laptop (so IO pool size 30). All with 100k rows per incomplete. All with 1000 incompletes.

## Memory analysis: With 10sec sleep on writes

Using debug build.

Without fix 3.7GiB at peak
10 concurrent workers 125 MiB at peak
100 concurrent workers 755 MiB at peak
1000 concurrent workers 3.7GiB at peak

## Performance Analyis with 3sec sleep on writes

Using release build.

Without fix 120s
10 concurrent workers 325s
60 concurrent workers (proposed default for my machine) 120s
100 concurrent workers 119s
1000 concurrent workers 121s
```

Also tested with a single IO and CPU thread and the semaphore only
having size 1 (to check we don't deadlock).