8066268338 Key segment pair refactor - do not move the segment out of the KeySegmentPair as it is a shared resource (#2166)
This allows us to re-enable some tests in `arcticdb-enterprise`
https://github.com/man-group/arcticdb-enterprise/pull/new/aseaton/keysegmentpair-refactor-enterprise
.

The changes are:

- Remove the mutable `Segment& segment()` in the `KeySegmentPair` - this
was often used to move out of the `Segment` owned by the shared pointer
in `KeySegmentPair`
- Pass `KeySegmentPair` by lvalue rather than rvalue reference where
possible. The `Task` layer still needs rvalue references as that is what
Folly expects (rightly, the Folly executors need to have ownership).
- Refactor `lookup_match_in_dedup_match` to return a variant
- Unrelated tidy up to `DecodeMetadataTask`
- `clone()` segments when adding them to the in in-memory and mock
storages

In particular this lets `CopyCompressedInterstore` safely copy the same
`KeySegmentPair` to several targets, whereas before the `Segment` could
be moved from during the copy to each target.