Fix off by one errors in staged segment validation (#2191)
Our validation when finalizing is too strict at the moment. The index
start and end on the APPEND_DATA key is `[start_time, end_time+1]` of
data contained within its segment. Our validation logic is not always
substracting one from the end time on the key to get the true date range
in the segment.

We need a similar change when detecting duplicate staged segments. If
two staged segments, both covering a single duplicated index value, are
staged, we should allow the write.