ASV framework refactor proposal
This commit adds some brief code example of how I think we can simplify
the current framework.

The main idea is that currently the `EnvConfigurationBase` is
responsible for too many things at the same time:
- dataframe generation
- library creation/population
- differentiating between persistant/modifiable libraries.
- managing different libraries needed for a benchmark

It being responsible for a large bundle of things makes it hard to:
- reuse between different benchmarks (we basically end up with one new
  child of `EnvConfigurationBase` for each benchmark)
- understand what some method actually does and we need to rely on
  lenghty documentation which is harder to keep up to date

My proposal is not code complete but show how we might go about
simplifying this framework. New code is inside the
`environment_setup_refactored.py` and show how we might go about
replacing the current `environment_setup.py`.

I've also tweaked the `LMDBReadWrite` benchmark to show how we can use
the new framework.

The proposal mainly consists of separating the four responsibilities of
the `EnvConfigurationBase`
- dataframe generation (now we have an interface `DataFrameGenerator`)
- library population (now we have an interface `LibraryPopulationPolicy`)
- persistant/modifiable (now we just have function which can give you
  what kind of library you need)
- managing different libraries (now this is entirely responsibility of
  the benchmark as each benchmark might need completely different logic
around the number/types of libraries it needs. E.g. read_write uses one
persistent and one modifiable)

By separating these we can get better reusablity:
- E.g. we can mix and match policies with dataframe generators for
  different benchmarks

Also now when I see the code in the benchmark I can how the libraries
are created/populated instead of an opaque `setup` call which is unclear
what it does.